Fix bug where empty section is sometimes shown
There is a bug where the empty section is
sometimes shown for additional providers. This
is because the logic for calculating the
visibility gets the providers and then the
logic for displaying the list gets the providers
and can result in getting out of sync. This aligns
both sets of logic to use the same data.

Test: ondevice & unit
Bug: 330163369
Change-Id: Icf71c78d67d55c929476f8e699d5893e60a49776

Former-commit-id: 1c3fb499133a6a12e59c17379b4e97913d4c7b09