Do not move to ALL_APPS when default Launcher is not system.
Currently we fire ALL_APP intent for System Launchers at the
end of PS setup flow, even when the system launcher (NexusLauncher)
was not the default Launcher.

In such cases, the user is moved to NexusLauncher, but they cannot
unlock PS (only default launcher can do it), leading to crashes.
To fix this, we do not fire the intent in case non system Launcher
has the HOME role.

Bug: 331722929
Test: Manual build and flash
Flag: NA
Change-Id: I77353474cc74028d72508c433c89b6116b4148bd

Former-commit-id: 63da8c35f7032912de22975882d4bf22c193d52a